Refactor name of public folder as 'old'
