--- public/styles/action_form.css
@@ -1,212 +1,0 @@
-.mn {
-  min-height: 48rem;
-  height: 80vh;
-}
-.mn--signup {
-  min-height: 67rem;
-  height: 100vh;
-}
-
-.hdng {
-  position: relative;
-  left: 50%;
-  display: inline-block;
-  transform: translateX(-50%);
-}
-.hdng-actn {
-  padding: 0.3rem 1rem;
-  float: left;
-  background: var(--denim, #1569ae);
-  border-radius: 1rem;
-  color: var(--white, #ffffff);
-  font-size: 5rem;
-}
-.hdng-side {
-  margin: 2rem 0 0 2rem;
-  float: left;
-  font-size: 2.5rem;
-}
-.hdng-link {
-  margin: 3rem 0 0 0;
-  position: relative;
-  left: 50%;
-  display: block;
-  float: left;
-  clear: left;
-  color: var(--denim, #1569ae);
-  font-size: 1.8rem;
-  transform: translateX(-50%);
-}
-
-.form {
-  margin: 5rem auto 0 auto;
-  padding: 3rem 1rem;
-  width: 100%;
-  max-width: 40rem;
-  background: #1569ae33;
-  border-radius: 1rem;
-}
-.form-lbl {
-  position: relative;
-  left: 16%;
-  display: block;
-  float: left;
-  clear: both;
-  font-size: 1.5rem;
-  font-weight: 700;
-  transform: translateX(-20%);
-}
-.form-lbl--signup_pswrd {
-  margin-top: 3rem;
-}
-.form-input {
-  margin: 0.5rem 0 1rem 0;
-  padding: 0.5rem 1.5rem;
-  width: 90%;
-  max-width: 32rem;
-  height: 3.8rem;
-  position: relative;
-  left: 50%;
-  display: block;
-  float: left;
-  clear: both;
-  border-radius: 1rem;
-  font-size: 1.6rem;
-  transform: translateX(-50%);
-}
-.form-input:hover,
-.form-input:focus,
-.form-input:focus-visible {
-  outline: 0.2rem solid var(--denim, #1569ae);
-}
-.form-err_lbl {
-  margin-bottom: 1rem;
-  width: 90%;
-  max-width: 32rem;
-  min-height: 2.5rem;
-  position: relative;
-  left: 10%;
-  display: block;
-  clear: both;
-  color: var(--rdornge, #ff3232);
-  font-size: 1.4rem;
-}
-.form-err_lbl:last-of-type {
-  margin-bottom: 0;
-}
-.form-err_lbl--hidden {
-  display: none;
-}
-.form-submit {
-  margin: 3rem 0 0 0;
-  width: 15rem;
-  height: 5rem;
-  position: relative;
-  left: 50%;
-  background: var(--white, #ffffff);
-  border-radius: 1rem;
-  color: var(--denim, #1569ae);
-  font-size: 2rem;
-  font-weight: 700;
-  transform: translateX(-50%);
-  transition: background ease-in-out 275ms, color ease-in-out 275ms;
-}
-.form-submit:hover,
-.form-submit:focus,
-.form-submit:focus-visible {
-  background: var(--denim, #1569ae);
-  color: var(--white, #ffffff);
-}
-
-@media screen and (min-height: 640px) {
-  .wrap--action_form {
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-}
-
-@media screen and (max-width: 800px) {
-  .mn--signup {
-    min-height: 63rem;
-  }
-  .hdng-actn {
-    padding: 0.3rem 0.7rem;
-    font-size: 4rem;
-  }
-  .hdng-side {
-    margin: 1.7rem 0 0 1rem;
-    font-size: 1.7rem;
-  }
-  .hdng-link {
-    margin: 2rem 0 0 0;
-    font-size: 1.6rem;
-  }
-}
-
-@media screen and (max-width: 550px) {
-  .mn {
-    min-height: 42rem;
-  }
-  .mn--signup {
-    min-height: 60rem;
-    height: 100vh;
-  }
-  .hdng-actn {
-    font-size: 2.5rem;
-  }
-  .hdng-side {
-    margin: 1rem 0 0 0.7rem;
-    font-size: 1.5rem;
-  }
-  .hdng-link {
-    margin: 2rem 0 0 0;
-    font-size: 1.4rem;
-  }
-  .form-lbl {
-    font-size: 1.4rem;
-  }
-  .form-err_lbl {
-    left: 8%;
-    font-size: 1.2rem;
-  }
-  .form-input {
-    font-size: 1.4rem;
-  }
-  .form-submit {
-    margin: 3rem 0 0 0;
-    width: 12rem;
-    height: 4rem;
-    font-size: 1.6rem;
-    font-weight: 700;
-  }
-}
-
-@media screen and (max-width: 320px) {
-  .mn {
-    min-height: 40rem;
-  }
-  .mn--signup {
-    min-height: 58rem;
-  }
-  .hdng-actn {
-    font-size: 2.2rem;
-  }
-  .hdng-side {
-    font-size: 1.4rem;
-  }
-  .hdng-link {
-    font-size: 1.2rem;
-  }
-  .form-lbl {
-    font-size: 1.3rem;
-  }
-  .form-input {
-    font-size: 1.3rem;
-  }
-  .form-submit {
-    width: 10rem;
-    height: 3.8rem;
-    font-size: 1.4rem;
-  }
-}
