+++ public_old/styles/global.css
@@ -1,0 +1,81 @@
+:root {
+  --denim: #1569ae;
+  --mnshft: #212121;
+  --white: #ffffff;
+  --wldsnd: #f4f4f4;
+  --grey: #8c8c8c;
+  --rdornge: #ff3232;
+}
+
+*,
+::before,
+::after {
+  margin: 0;
+  padding: 0;
+  border: none;
+  box-sizing: border-box;
+  font-family: "Roboto", Arial, Helvetica, sans-serif;
+  color: var(--mnshft, #212121);
+}
+
+html {
+  font-size: 10px;
+}
+
+.wrap {
+  width: 100%;
+  position: relative;
+}
+
+.hdr {
+  padding: 0 min(4rem, 5vw);
+  position: relative;
+  height: 8rem;
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: flex-end;
+  align-items: center;
+}
+.hdr-log {
+  width: 10rem;
+  height: 4rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: transparent;
+  border-radius: 1rem;
+  color: var(--denim, #1569ae);
+  text-decoration: none;
+  font-size: 1.8rem;
+  font-weight: 700;
+  transition: background ease-in-out 275ms, color ease-in-out 275ms;
+}
+.hdr-log:hover,
+.hdr-log:focus,
+.hdr-log:focus-visible {
+  background: var(--denim, #1569ae);
+  color: var(--white, #ffffff);
+}
+
+.mn {
+  margin: 0 auto;
+  width: min(80%, 100rem);
+}
+
+.ftr {
+  position: relative;
+  height: 6rem;
+}
+.cprgt {
+  padding: 3rem 1rem;
+  width: min(40rem, 90%);
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  text-align: center;
+  font-size: 1.2rem;
+}
+.cprgt-link {
+  color: var(--denim, #1569ae);
+}
